rename auth0 to auth
