+++ bcc-orgs/src/auth/verify.go
@@ -1,4 +1,4 @@
-package auth0
+package auth
 
 import (
 	"bcc-orgs/src/utilities"
